requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.0.3
spacy==3.7.2
streamlit==1.28.0
sentence-transformers==2.2.2
